Add files via upload
A Java program that repeatedly takes the details of a mortgage from the user and calculates the required monthly repayment.  In addition, it asks the user how much they can afford to pay per month and calculate how long it would take to repay the mortgage. It assumes that the mortgage is 20 years in duration.