Add files via upload
A Java program which uses the Sieve of Eratosthenes to compute the prime numbers from 2 up to a given number N which is entered by the user. As part of your solution you must use an array to represent the sequence of numbers (which may contain both crossed out and non-crossed out numbers).